requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.7.2,<3
pydantic-settings>=2.5.2
email-validator==2.1.0
python-multipart>=0.0.9
aiofiles==23.2.1
duckdb==0.9.2
qdrant-client==1.12.0
sentence-transformers>=2.7.0
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
unstructured[all-docs]==0.14.9
beautifulsoup4>=4.12.0
python-dotenv==1.0.0
httpx>=0.27.1,<0.28
cachetools==5.3.2
pandas==2.1.4
pyarrow>=14.0.0
openpyxl==3.1.2
xlrd>=2.0.1
python-magic==0.4.27
nltk==3.8.1
matplotlib==3.8.2
psutil==5.9.8
pdfminer.six==20221105
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# LLM Providers (Phase 3.V.3)
google-generativeai>=0.8.0
google-genai>=1.0.0,<2
openai>=1.6.1,<1.68.0
anthropic>=0.42.0

# Prompt Templates (Phase 3.V.4)
jinja2>=3.1.0

# State Management (Phase 3.W.1)
sqlmodel>=0.0.14

# Database Migrations (BQ-111: Persistent State)
alembic>=1.13.0

# Encryption (BQ-066: API key encryption at rest)
cryptography>=42.0.0
psycopg2-binary==2.9.9

# Observability (BQ-123A)
structlog>=24.1.0
pyyaml>=6.0

# Local Auth (BQ-127: Air-Gap Architecture)
bcrypt>=4.0.0

# Multi-User Auth (BQ-VZ-MULTI-USER: Admin/User Role Split)
argon2-cffi>=23.1.0
PyJWT>=2.8.0

# MCP Server (BQ-MCP-RAG: Universal LLM Connectivity)
mcp>=1.8.0,<1.9

# SQL AST validation (BQ-MCP-RAG: defense-in-depth)
sqlglot>=20.0.0

# BQ-VZ-DB-CONNECT: External database connectivity
sqlalchemy>=2.0
pymysql>=1.1

# BQ-VZ-LARGE-FILES: Streaming PDF processing
pypdfium2>=4.0.0
pdfplumber>=0.10.0
ijson>=3.2.0

# BQ-VZ-AUTO-UPDATE: Docker SDK for auto-update via Docker socket
docker>=7.0.0
websockets>=14.0,<15.0

# BQ-VZ-PERF Phase 3: Native document handlers (replace Tika sidecar)
striprtf>=0.0.26
icalendar>=5.0.0
vobject>=0.9.6
ebooklib>=0.18
odfpy>=1.4.1
extract-msg>=0.48.0
feedparser>=6.0.0
